+++ clang-tools-extra/clang-tidy/hicpp/SignedBitwiseCheck.cpp
@@ -36,7 +36,8 @@ void SignedBitwiseCheck::registerMatchers(MatchFinder *Finder) {
   Finder->addMatcher(
       binaryOperator(
           allOf(anyOf(hasOperatorName("^"), hasOperatorName("|"),
-                      hasOperatorName("&")),
+                      hasOperatorName("&"), hasOperatorName("^="),
+                      hasOperatorName("|="), hasOperatorName("&=")),
 
                 unless(allOf(hasLHS(IsStdBitmask), hasRHS(IsStdBitmask))),
 
@@ -48,10 +49,11 @@ void SignedBitwiseCheck::registerMatchers(MatchFinder *Finder) {
   // Shifting and complement is not allowed for any signed integer type because
   // the sign bit may corrupt the result.
   Finder->addMatcher(
-      binaryOperator(allOf(anyOf(hasOperatorName("<<"), hasOperatorName(">>")),
-                           hasEitherOperand(SignedIntegerOperand),
-                           hasLHS(hasType(isInteger())),
-                           hasRHS(hasType(isInteger()))))
+      binaryOperator(
+          allOf(anyOf(hasOperatorName("<<"), hasOperatorName(">>"),
+                      hasOperatorName("<<="), hasOperatorName(">>=")),
+                hasEitherOperand(SignedIntegerOperand),
+                hasLHS(hasType(isInteger())), hasRHS(hasType(isInteger()))))
           .bind("binary-sign-interference"),
       this);
 
@@ -84,10 +86,8 @@ void SignedBitwiseCheck::check(const MatchFinder::MatchResult &Result) {
     else
       llvm_unreachable("unexpected matcher result");
   }
-
-  diag(Location,
-       "use of a signed integer operand with a %select{binary|unary}0 bitwise "
-       "operator")
+  diag(Location, "use of a signed integer operand with a "
+                 "%select{binary|unary}0 bitwise operator")
       << IsUnary << SignedOperand->getSourceRange();
 }